requirements.txt
streamlit
openai
pandas
requests
datetime
plotly.express